fix issues, update comments
