+++ mdk/utils.go
@@ -4,7 +4,11 @@ package mdk
 // #include <string.h>
 import "C"
 import (
+	"fmt"
 	"unsafe"
+
+	"github.com/wasify-io/wasify-go/internal/types"
+	"github.com/wasify-io/wasify-go/internal/utils"
 )
 
 // free deallocates the memory previously allocated by a call to malloc (C.malloc).
@@ -83,4 +87,12 @@ func stringToLeakedPtr(data string, offsetSize uint32) (offset uint64) {
 // return a pointer of type T pointing to the data at the specified memory address.
 func ptrToData[T any](ptr uint64) *T {
 	return (*T)(unsafe.Pointer(uintptr(ptr)))
+}
+
+func unpackDataAndCheckType(packedData ArgData, expectedType types.ValueType) (types.ValueType, uint32, uint32) {
+	valueType, offsetU32, size := utils.UnpackUI64(uint64(packedData))
+	if valueType != expectedType {
+		LogError(fmt.Sprintf("Unexpected data type. Expected %d, but got %d", expectedType, valueType))
+	}
+	return valueType, offsetU32, size
 }
